project-ahaa/requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.2
pinecone>=8.0.0
pypdf>=3.17.0
python-docx>=1.1.0
nltk>=3.8.1
pandas>=2.1.0
networkx>=3.2
pyvis>=0.3.2
numpy>=1.24.0
